+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 -   `EncryptionSettings.onlyAllowTrustedDevices` has been replaced with
     `EncryptionSettings.sharingStrategy`, which adds the ability to share only
     with cross-signed devices.
+    ([#134](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/134))
 
 **Other changes**
 
--   Update matrix-rust-sdk to `11cbf849c`, which includes:
+-   Add `OlmMachine.registerRoomKeysWithheldCallback` to notify when we are
+    told that room keys have been withheld.
+    ([#136](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/136))
+
+-   Update matrix-rust-sdk to `8d54bd92d`, which includes:
 
     -   refactor(sdk-crypto): Room key sharing, introduce extensible strategy
         ([#3605](https://github.com/matrix-org/matrix-rust-sdk/pull/3605))